moved name-concat to own component
