+++ src/app/app.module.ts
@@ -9,6 +9,7 @@ import { HelloComponent } from './hello.component';
 import { AppRoutingModule } from './app-routing.module';
 import { CommonModule } from '@angular/common';
 import { BasicFormComponent } from './basic-form/basic-form.component';
+import { NameConcatComponent } from './name-concat/name-concat.component';
 
 @NgModule({
   imports: [
@@ -20,7 +21,12 @@ import { BasicFormComponent } from './basic-form/basic-form.component';
     BrowserAnimationsModule, // required animations module
     ToastrModule.forRoot(), // ToastrModule added
   ],
-  declarations: [AppComponent, HelloComponent, BasicFormComponent],
+  declarations: [
+    AppComponent,
+    HelloComponent,
+    BasicFormComponent,
+    NameConcatComponent,
+  ],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
