+++ src/app/basic-form/basic-form.component.html
@@ -109,63 +109,8 @@
 </div>
 
 <!-- Name Concat -->
-<div class="container p-5">
-  <h1 class="p-2">Name Concatenation</h1>
-  <form [formGroup]="nameForm">
-    <!-- FIRST NAME -->
-    <div class="row mb-2">
-      <label label for="firstname" class="col-4 col-form-label"
-        >First Name</label
-      >
-      <div class="col-8">
-        <input
-          type="text"
-          class="form-control"
-          id="firstname"
-          formControlName="firstname"
-          placeholder="First Name"
-        />
-      </div>
-    </div>
-    <!-- LAST NAME -->
-    <div class="row mb-2">
-      <label label for="lastname" class="col-4 col-form-label">Last Name</label>
-      <div class="col-8">
-        <input
-          type="text"
-          class="form-control"
-          id="lastname"
-          formControlName="lastname"
-          placeholder="Last Name"
-        />
-      </div>
-    </div>
-    <!-- OUTPUT -->
-    <div class="row mb-2">
-      <label label for="nameoutput" class="col-4 col-form-label">Output</label>
-      <div class="col-8">
-        <input
-          type="text"
-          class="form-control"
-          id="nameoutput"
-          placeholder="Result"
-          value="{{ resNameConcat }}"
-          readonly
-        />
-      </div>
-    </div>
-    <button (click)="onNameClick()" class="col-5 btn btn-primary">
-      Join Names
-    </button>
-  </form>
+<app-name-concat></app-name-concat>
 
-  <div>
-    <p>Name Value: {{ nameForm.value | json }}</p>
-    <p>Name Valid: {{ nameForm.valid | json }}</p>
-    <p>Firstname Error: {{ firstname.errors | json }}</p>
-    <p>LastName Error: {{ lastname.errors | json }}</p>
-  </div>
-</div>
 
 <!-- BMI Calc -->
 <div class="container p-5">
